windows: Overwrite DEFAULT_CONFIG_DIR on Windows
Paths of form: "/opt/mesosphere/etc/dcos-net.config.d" do not exist
on Windows, so we switch to a environmet define for it.

This patch overwrites the pre-existing define in case `WINDOWS` define is
set.

Also define `WINDOWS` in erl_opts from rebar. This may be useful in the long
run.

Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>